Backend/requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
fastapi==0.61.2
h11==0.11.0
idna==2.10
Pillow==8.0.1
pydantic==1.7.2
requests==2.25.0
starlette==0.13.6
todoist-python==8.1.2
typing==3.7.4.3
urllib3==1.26.2
uvicorn==0.12.3
